# docs/swagger.yaml
basePath: /
definitions:
...
  dto.GetAllMedicinesResponseDto:
    properties:
      medicines:
        items:
          $ref: '#/definitions/dto.MedicineResponseDto'
        type: array
      total:
        type: integer
    type: object
...
  dto.GetMedicineByIDResponseDto:
    properties:
      medicine:
        $ref: '#/definitions/dto.MedicineResponseDto'
    type: object
...
  dto.MedicineResponseDto:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: number
      unit:
        type: string
      updated_at:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: localhost:8082
info:
  contact: {}
  description: This is a sample server for a user API.
  title: Order API
  version: "1.0"
paths:
...
  /api/medicine/v1/medicines:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all available medicines in the system. This
        endpoint does not require authentication.
      produces:
      - application/json
      responses:
        "200":
          description: Medicines retrieved successfully
          schema:
            $ref: '#/definitions/dto.GetAllMedicinesResponseDto'
        "500":
          description: Internal server error while retrieving medicines
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get all available medicines
      tags:
      - medicines
  /api/medicine/v1/medicines/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves detailed information about a specific medicine identified
        by its ID. This endpoint does not require authentication.
      parameters:
      - description: Medicine ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Medicine retrieved successfully
          schema:
            $ref: '#/definitions/dto.GetMedicineByIDResponseDto'
        "400":
          description: Invalid or missing medicine ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Medicine not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error while retrieving medicine
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get medicine details by ID
      tags:
      - medicines
...
schemes:
- http
swagger: "2.0"